HW/D/08-aurora: address race between kernel starts
The strm_dump kernel needs to be up and running when the strm_issue
kernel starts to transfer data across the Aurora channel. Otherwise,
data will be lost and the application will deadlock, waiting for the
correct amount of data on the receiver side.

In practice we observe that calling run_strm_issue.start() directly
after run_strm_dump.start() does not guarantee that the strm_dump kernel
is running early enough. Introducing a small delay of 1ms is enough to
make it work reliably.