HW/D/08-aurora: use correct refclk
The Aurora core is configured to be fed with a clock of 161.1328125 MHz.
However, on some Alveo boards it is currently fed with a 156.25 MHz
clock. On those, refclka needs to be replaced by refclkb, which actually
represents the 161.1328125 MHz clock.

This change directly affects the throughput (measured on a U280):

Before: 4594.95 MB/s
After:  4733.95 MB/s

The latter result is closer to the expected 40 Gbps (5000 MB/s).